requirements.txt
pandas
requests
websocket-client
configparser